Add ephemeral variants for nonephemeral runners (#6377)
With the goal of moving away from nonephemeral runners, we plan to
proceed this migration workflow-by-workflow.

So the goal is to create a `ephemeral` variant of the current
nonephemeral runners, and eventually migrate all workflows to that
runner. Once successful, migrate back the workflows to the original
label, marking that label as ephemeral by default.