Adds additional tests to getRunnerTypes, simplifies code a bit, adds the support for `c.` runners (#6403)
## Support for `c.` runners

We don't really support well our canary environment with variants, not
sure why, but this condition was not present, so variants will get the
wrong naming when running at meta canary environment.
`variant.c.runner.name` instead of `c.variant.runner.name`. This fix is
very low impact and should not change anything in production.

## Simplifies the code

There is a useless `if` that I noticed when reading `getRunnerTypes`. I
am removing that conditional;

## Additional tests for getRunnerTypes

As we're adding empty variants as part of the ephemeral migration, I
want to make sure that we have green tests and we're covering this
situation in our unit tests.

The new situations to cover are:
1) Empty variants
2) meta's canary environment naming convention `c.`