Add ephemeral variants for all runner types, and updates validate_scale_config.py to enforce it (#6402)
# TLDR

Adds `ephemeral` variant to all runner types, and updates
validate_scale_config.py to enforce it

# What?

Enable experiment with ephemeral runners for all workflows. The status
of the experiment can be found in the [test-infra
issue](https://github.com/pytorch/test-infra/issues/5132).

# Why?

Those runners are ephemeral, as eliminating nonephemeral runners is a
follow up for the recent security incident. Refreshable infrastructure
have been something we've trying to accomplish for a while, but haven't
been successful. The major blocker we face is related to stockouts and
unreliability from GH side. Most of it is because nonephemeral runners
can run other jobs and continue clearing the queue in case of a problem.
This is not possible for ephemeral runners.

# How?

To remediate stockouts, the [reuse/refresh of ephemeral
instances](https://github.com/pytorch/test-infra/pull/6315) have been
introduced.

In order to remediate GH side issues, [queue healing
mechanism](https://github.com/pytorch/test-infra/pull/6390) is being
implemented.

Also, in order to guarantee the stability of this behaviour, [additional
unit tests have been included with other
changes](https://github.com/pytorch/test-infra/pull/6403).

# Next steps

After merging those changes, we intend to put a small percentage of jobs
to use ephemeral runners, so we can evaluate impact on queue times and
gather statistics on reuse and tune noflap cluster behaviour. Once we
feel comfortable the experiment will be shifted to 100% and we'll
migrate all workflows to fully ephemeral instances. Eventually, all
runners will be ephemeral and the experiment and runner variants will be
removed, as we update other workflows like slow, trunk and nightlies.

# The ephemeral runners are not working properly?

Then go to the experiment on the `What?` section of this document and
set the experiment named `ephemeral` it to 0.