Remove Lf metrics from Metrics page on hud (#6413)
As we finished the migration to support autoscaled runners in LF and
we're now comfortable in a stable situation for a few months, the
following graphs in the Hud metrics page are not relevant anymore:

<img width="619" alt="Screenshot 2025-03-14 at 18 47 13"
src="https://github.com/user-attachments/assets/4662540e-1fc5-4ce0-b631-72110ab91644"
/>

This change is removing them from that page, and the queries that are
used by it.

The dashboard that displays the % of usage in each fleet is still very
relevant, and we'll be keeping it.

# Release and rollback

There is CI in this repository that triggers the release every time
changes are merged to main. So, both for release and rollback just merge
to main and wait changes to be reflected in hud.pytorch.org.