[HUD] Adds dashboard in Metrics page to track ephemeral experimentation % over time (#6420)
Adds a dashboard with title "Percentage of jobs on ephemeral experiment"
on metrics page of hud.

Query is a copy-paste-change-a-bit from meta/lf runners. This dashboard
is temporary and we should be removing it as we finish the experiment
and the migration.

Once we do that, the above mentioned dashboard and the `nonephemeral`
metric in SLI page should be removed.