[BE][drci] Async fetch PR info during reorganizeWorkflows for speed (#6380)
Not sure if github is good enough at showing the diff here since the
highlighting what is indentation is helpful for seeing what actually
changed. My vscode changed files is more helpful

Instead of doing a synchronous for loop, use async to reduce the time
needed to wait for api calls.

This reduces the time this function takes from ~45s -> 8s (obviously
depends on how many PRs, but I assume this is pretty representative of
calling drci on pytorch/pytorch during the workday). Waiting for the 45s
is kind of annoying when trying to test changes to dr ci

I tested that the result of this didn't change by comparing the results
of of the output of the function