Adds linux.4xlarge.for.testing.donotuse for testing (#6461)
Adds the runner `linux.4xlarge.for.testing.donotuse` that will be used
to test scaleUpChron in production.

The goal is to deploy version of scaleUp lambda that ignores this label,
and open a pytorch PR requesting this label. This job will then be
queued and never be picked up, until (hopefully) scaleUpChron lambda
deploys a runner for it.

After the testing, we can revert this commit.