[BE] Remove duplicate, stale copy of github_analyze script (#6490)
There were two versions of this file, one an old fork of the other from
back when pytorch/builder existed. Deleting the older forked version of
this file in favor of the updated analytics version. The fork had
exactly one significant change (an extra parameter that it accepted), so
integrated that change into the script we're preserving

github_analyze.py is called in exactly two situations: by reverts.py
(it's import has been updated) and manually by github_analyze.py