[bots] Allow flakybot to make aggregate issues (#6649)
Requires changes in https://github.com/pytorch/test-infra/pull/6661,
please review + merge that one first

Allow flakybot to make aggregate issues, update them, and close them.
Also allow the verification bot that comments on each issue to do
aggregate issues.

Format should match the one in
https://github.com/pytorch/test-infra/pull/6045

We should really make the verification script just the same as the
script that updates the json...

No platform module labels are going to be added

There's some weirdness with closing issues / removing tests that are no
longer flaky because I think it's hard (read: possible, probably
annoying) to determine which test is responsible for an issue getting
updated and which are old enough to be removed

Threshold is 10