[bots] Move the disable json generation to be in HUD (#6676)
We get the list of disabled/unstable test/jobs in a python script, but
there is code that parses the same thing in javascript for the bot that
comments on all disable/unstable test/job issues, so this removes the
python script in favor of a new javascript script that reuses the code
used for verifyDisableIssue to generate the disable/unstable test/job
jsons.

I broke the unstable/disable job json generation in
https://github.com/pytorch/test-infra/pull/6096, so it's actually been
empty since then. We've been getting by, and this PR restores it, but I
wonder if this just means we can get rid of it

Pros:
* less dup code
  * string parsing for title to decide which type of issue it is
* string parsing for body to figure out what the platforms are/what the
tests are

Cons:
* it wasn't broken
* stuff that wasn't dup and had to be rewritten:
* generating json objects of the correct shape; merging issues/platforms
for the same test if necessary
  * querying graphql for list of issues
  * some parsing for unstable/disable job issues

CI is broken on this PR because it calls an API that this PR adds (which
basically means this isn't really testable in CI anymore... but also
given that I broke some things months ago and didn't notice, the e2e for
this is probably not great in the first place). I tested locally with
localhost and with the preview and it looks fine to me