runner: Double check for runnerMinimumTimeExceeded (#6703)
We found issues / instances where runnerMinimumTimeExceeded would make
decisions with outdated data based on if it had recently undergone a
root volume replacement task since scale-down is exceeding ~8 minutes
per execution.

Solution here is to just run runnerMinimumTimeExceeded twice one with
cached data and one with fresh data to ensure that we are doubly sure
that we should be spinning something down.

---------

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
Co-authored-by: Zain Rizvi <ZainRizvi@users.noreply.github.com>