Add setup-python action (#6728)
So far only works on MacOS and uses homebrew. Must be implemented as
JScript action, otherwise one can not have a cleanup step.

Needed to get rid of Anaconda dependency for MacOS builds and tests
Tested by test-setup-python.yaml, that is triggered on modifications to
all 4 files in question as well as
https://github.com/pytorch/pytorch/pull/155515

Partially vibe-coded in
https://chatgpt.com/s/cd_684780ca34108191973cd18bfd22f8a7 but than had
to do some fixes/cleanups afterwards