[Wiz Sensor] Enable runners to have specific secrets be made available to them (#6739)
There's a secret we want to enable runners to read

We add two new variable to the module to specify:
- Arn for the secret we want the runner to read (stored in secret
store).
- Arn for the decryption key (stored in kms).

The variables are marked as sensitive in order to not show up in the
logs, and the runner is then given a policy allowing it to access the
secret directly from secret store.

More notes:
- The policy is only conditionally added when a secret is actually
passed in. Not all the fleets will use this secret, so we will need to
make sure to handle the secret not being available in the runner logic.
- The secret's arn is the secret key plus a random suffix. We use a wild
card to account for the suffix and allow people to pass in bare secret
names (without the random suffix, which could change if the secret gets
deleted and recreated)

Testing: Manually applied the policy in the Meta AWS account and
verified that I could read the secret from a runner instance