Install wiz sensor on linux runners (#6750)
Installs Wiz Runtime Sensor on Linux runners during startup, building on
https://github.com/pytorch/test-infra/pull/6739/. This adds a layer of
security to our CI jobs and opens up a pathway for code scanning as
well.

The wiz runtime sensor requires credentials that it gets from a secret
who's name is passed in to the terraform module name. For environments
where that credential is not available, the installation is skipped.

Why only linux? Windows isn't yet supported by the wiz sensor (but is on
the way)

## Graceful failure
Since it's important to make sure the runners can come online even if
the wiz runtime sensor installation isn't possible, this PR contains
**multiple graceful failure protocols** to allow the runner to come
online even if the agent cannot be installed:
1. If the secret arn was never passed into the terraform module (which
is sometimes expected), the user-data.sh script doesn't even include the
code to install wiz sensor.
1. If the secret passed in was invalid, then there are multiple levels
of checks to skip the installation and continue regular runner setup
1. If the secretsmanager is unable to get the secret value for any
reason, skip the installation and continue regular runner setup
1. If the sensor installation fails (e.g. the wiz server hosting the
code is down), skip the installation and continue regular runner setup

All the unexpected failure modes result in metrics being emitted.

## Testing
Ran code on an EC2 runner and verified it resulted in wiz sensor running
and connecting to the wiz backend as expected. Manually tested both
happy path and failure modes

## Follow up
This needs to be followed up by one more PR in gha-infra to actually
pass in the secret arn to the terraform module