aws: Add tool to cleanup-ssm parameters
This is supposed to eventually deprecate our ssm parameter cleanup in
scale-down.ts.

First step here is to just have the utility, it's got unit tests and
integration tests. I've also gone ahead and ran it myself a couple of
times to double check on how it is.

We might want to add functionality to read config values from environment
variables but this is a first sweep at adding this.

Also adds some documentation as well. Mostly vibe-coded but reviewed by
seemethere

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: 38e2d27c9d779e1dff3bebdd5b5e0f39a475ee83
ghstack-comment-id: 2973329902
Pull-Request: https://github.com/pytorch/test-infra/pull/6770
Signed-off-by: Eli Uriegas <eliuriegas@meta.com>