aws: Add tool to cleanup-ssm parameters (#6770)
This is supposed to eventually deprecate our ssm parameter cleanup in
scale-down.ts.

First step here is to just have the utility, it's got unit tests and
integration tests. I've also gone ahead and ran it myself a couple of
times to double check on how it is.

We might want to add functionality to read config values from
environment
variables but this is a first sweep at adding this.

Also adds some documentation as well. Mostly vibe-coded but reviewed by
@seemethere

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>