Keep going display on HUD: log classifer to handle temp logs (#6723)
Context is https://github.com/pytorch/pytorch/pull/155371

Add a temp_log input to read the log from the temp_log bucket and put
the attribute into a temp attribute

This should be a noop in the normal flow

I'm not actually sure if this will work, since the webhook to dynamo
code seems to overwrite the entire dynamo entry, so if another webhook
gets sent, this attribute would be lost. But webhooks only get sent when
something happens, so maybe only when a step finishes or the job
finishes? so it might be ok. But also the normal flow doesn't use this
so it should be safe in that sense to merge

I'm also not sure if a temp attribute is necessary, since if the above
happens the temp attribute will be gone

Possible solutions if the above happens and the temp attribute
disappears:
* Insert into different table/s3 bucket -> different clickhouse table ->
do a join
  * I think this will have bad perf b/c join
* Change webhook to dynamo to be update instead of overwrite
  * not sure if the fields of the webhook change

Testing: in fixtures/request.json, changed the raw query string to
`job_id=44069312114&repo=pytorch%2Fpytorch&temp_log=true` and did the
things in the readme and saw that the temp value got written to dynamo
under a new temp attribute, also edited the clickhouse tables to add
columns and saw that it got inserted. Also ran it with without temp_log
and saw that the normal log got used.