Enable specifying whether or not to pull the docker image from ECR (#6836)
We originally tried to simplify the number of parameters users had to
pass in to linux_job_v2 by using the existence of the docker build
script as a heuristic.

That worked for a while, but that heuristic is now starting to break. If
a repo has an unrelated docker build script located in
`.ci/docker/build.sh`, we assume it wants to pull from our custom ECR
path instead of docker hub.

Our ideal end state would now look like:

1. Have an explicit parameter that needs to be set if a repo wants to
use our ECR registries instead of assuming the desire based on the
existence of a specific file. We default to docker hub
2. Migration to this end state would include fix all existing domain
repos that have that docker build file defined to pass in that
parameter.

That migration requires many repos to be edited with a fix however, so
for now we're having the following non-breaking change:

1. Have an explicit parameter on linux_job_v2 that needs to be set if a
repo wants to use our ECR registries. *Set it to TRUE by default*, with
the calculate-docker-image logic being "If param is set to True AND
special docker file exists" then we use the ECR registry. If either is
false, use docker hub

2. No migrations needed in the short term. We can later migrate domain
repos to explicitly set this setting and then change the default to
FALSE, but that's P2

Testing: Ran [this
job](https://github.com/pytorch/pytorch/actions/runs/15884049575/job/44791525110)
against it, and it only failed due to an error in the script inputted to
the workflow, which is ignorable