runners: Add terraform module for scale-cycle
Adds terraform files required for deploying scale-cycle. This also adds
a switch to turn on scale-cycle or not with the default position being
off. This should make it easier to deploy scale-cycle in the future and
could also act as a gate for easily enabling / disabling scale-up's
tryReuseRunner functionality.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: 87583f4ca794b515b1a8f176c476d41c977cbe19
ghstack-comment-id: 3046573503
Pull-Request: https://github.com/pytorch/test-infra/pull/6893
Signed-off-by: Eli Uriegas <eliuriegas@meta.com>