runners: Move runner removal logic up
This is a refactor of scale-down to move the runner removal logic up to the top of the loop.
This is done to avoid long wait times between determining if a runner should be removed and actually removing it.

In practice we were observing wait times of up to 7 to 10 minutes.

This might only actually be testable with production traffic / rate limits.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>