runners: Add expiration policy to SSM parameters (#6885)
This adds an expiration policy to the SSM parameters for the runners.
This is to ensure that the parameters are deleted after 30 minutes.

Github Runner Tokens typically have a 1 hour expiration time, but our
runners are typically expected to be up way quicker than that so 30
minutes is a good balance for when we expect the runners to be up.

If a runner isn't conencted to Github by at least 30 minutes we will
more than likely have spun it down and it will be deleted.

This is an attempted re-land of 2 commits:
* #6855
* #6858

---------

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>