[autorevert] When checking job status, only expect target job list to be finished, not waiting for all jobs to have finished (#7023)
There are some pending jobs checks that are overly cautious and could be
problematic in case of queuing of some exotic type of runner. This can
hold autorevert hostage if there is a long queue or an outage in some
type of runner that is unrelated to the patterns we're looking for.

## Older commits (baseline)

When checking for patterns, we don't really care if there are pending
jobs for the older commit, as long it is unrelated to the identified
pattern.

## Repeated error identification

Already correctly handled, it should retry jobs at first opportunity of
a single repeated breaking job.

## `confirm_commit_caused_failure_on_restarted`

`has_rule` already checks for failure conclusion. So it is more correct
ignore if jobs are still pending and match in the first confirmation.
But it is important to only check for the relevant job names if they
finished in base, as any newly job could finish as failure, what should
avoid reverting.

---------

Co-authored-by: Ivan Zaitsev <ivanzaitsev@fb.com>