[BE] Cleanup the old torchao benchmark dashboard from TorchBench (#7034)
This is definitely not coming back. We can reuse the name of TorchAO
dashboard on the benchmark running on torchao itself. Removing the old
dashboard from HUD navigation bar is a good start, and there are more
things to clean up here which I will follow later in a separate PR.

Signed-off-by: Huy Do <huydhn@gmail.com>