Fixing the behaviour for getRunnerTypes with scaleConfigOrg (#7062)
Currently, when scaleConfigOrg is pointing to an organization that is
not the one runners are assigned to, it is not always correctly selected
for `getRunnerTypes` call.

Triggering errors similar to:

```
ERROR [getRunnerTypes]: HttpError: Not Found
```

This is due it not be correctly matched in all places where its usage is
called.