[autoscalers] Only use auth to download github files if needed (#7064)
This enables autoscalers to use a scale-config that's located in an
organization other than the one they're located in, as long as that
scale-config is located in a public repo (which all our scale configs
currently are).

Bug it fixes: The old code would create a github client to download the
scale-config.yml file, but `createGitHubClientForRunnerOrg` will fail if
you try to try to create a client for an org your app doesn't have
access to. Using a full blown git client for a public file also seems
unnecessary.

This version uses a normal http request to pull the raw file. So
authentication doesn't matter. (Aside: I considered keeping the old flow
as a backup path for if we ever want the scale config to live in a
public repo, but if and when that day comes I'd rather we add the logic
afresh than leave dead, unused code around in the script.)

Testing: Verified the getRunnerTypes functionality locally to ensure it
worked end-to-end without mocks.

---------

Co-authored-by: Jean Schmidt <4520845+jeanschmidt@users.noreply.github.com>