[HUD] Update some queries and regex for better viable/strict job definitions (#7093)
Similar to https://github.com/pytorch/pytorch/pull/162100

Reason:
rocm binary builds should not be included in viable strict upgrades and
metrics that look at viable strict. Also add linux-aarch64 to some since
that was missing

Is this the correct way to measure trunk redness? Idk but I want it be
consistent with what it says it is (considers only viable/strict
blocking jobs) so we don't get confused

Some have spaces after them b/c thats the easier way to guarantee that
we won't accidentally include linux-binary-manywheel-rocm when we do
regex for linux-binary-manywheel, and convention states that job names
that include workflow is `workflow name / job name`. The places we don't
need it are places that only look at workflow name and look for strict
equality