Upgrade runner-binaries-syncer to aws sdk v3 (#7078)
`yarn test` is broken on main, you can see this in
https://github.com/pytorch/test-infra/blob/a32b8f647ed2df0e93a167e518cf92f5855671ce/terraform-aws-github-runner/modules/runner-binaries-syncer/lambdas/runner-binaries-syncer/Makefile#L15


Testing:
Stole some environment variables from the lambda and mangled the key to
upload to a dummy key then ran `yarn build; cd dist;node -e
'require("./index").handler();' > t.log`
Saw that it uploaded a file, and skipped some because they didn't need
to be uploaded.
The one that was uploaded was arm64, which I'm thinking was uploaded
manually since it lacks a tag on s3.

I had to add an `await` since it wasn't working, which I think is a bug
in the original code