[autorevert] Breaks down dry_run logic for revert and restart (#7179)
*PLEASE NOTE*
this PR is intended to be on top of
https://github.com/pytorch/test-infra/pull/7179
maybe merge that one before reviewing this one to make reviewing easier

Breaking down the dry_run logic for revert and restart is required so we
can continue to safely work towards improving the autorevert safely.

Not adding the logic is not the best option, as I hope we'll be able to
run it locally a few times and iterate on that before publishing. And
rely on code commenting and other not so great approaches is not ideal.

---------

Co-authored-by: Ivan Zaitsev <ivanzaitsev@fb.com>