[ALI] adds AUTH_GH_ORG and AUTH_GH_REPO to set/restrict authentication and management for a specific org/repo (#7152)
Currently we've been guessing org/repo to authenticate in most cases,
this configuration when set both enforce validation and defaults to the
given org/repo for authentication.

This enables scaleUpChron to run properly when the org/repo for the
scale config is not the same as the one used to authenticate and being
managed.

additional small fixes to remove warnings on linting that have been
ignored.