[BE] Split up download.pytorch.org index update job (#7185)
The jobs is getting slower and close to 1 hours now
https://github.com/pytorch/test-infra/actions/workflows/update-s3-html.yml
and I hate waiting an hour for the index to update. It's also awkward
trying to explain to folks that they need to wait an hour for the index
to update after a package is uploaded.

My proposal is to split up the job and run it on more runners for each
accelerators `cuda`, `rocm`, `xpu`, `cpu`.

I also accidentally add a local file to the folder in my previous PR
https://github.com/pytorch/test-infra/pull/7182 here.

### Testing

https://github.com/pytorch/test-infra/actions/runs/17815392432

---------

Signed-off-by: Huy Do <huydhn@gmail.com>